refactor: show several types of signatures
