+++ webapp/src/lib/schema/CRUDForm.svelte
@@ -74,7 +74,9 @@
 		...formSettings
 	};
 
-	const dispatch = createEventDispatcher<{ success: {} }>();
+	const dispatch = createEventDispatcher<{ success: {
+		record: PBRecord;
+	} }>();
 
 	/* Schema generation */
 
@@ -99,12 +101,13 @@
 			async ({ form }) => {
 				const data = cleanFormDataFiles(form.data, fileFieldsInitialData);
 				const formData = createFormData(data);
+				let rc;
 				if (mode == formMode.EDIT && initialData && initialData.id) {
-					await pb.collection(collection).update(initialData.id, formData);
+					rc = await pb.collection(collection).update(initialData.id, formData);
 				} else {
-					await pb.collection(collection).create(formData);
+					rc = await pb.collection(collection).create(formData);
 				}
-				dispatch('success', {});
+				dispatch('success', {record:rc});
 			},
 			mockedData
 		);
