+++ webapp/src/routes/my/signatures/+page.svelte
@@ -10,7 +10,6 @@
 		createSlotTypeCaster
 	} from '$lib/schema/recordsManager/recordsManager.svelte';
 	import RecordsTable from '$lib/schema/recordsManager/views/recordsTable.svelte';
-	import SignDocumentButton from '$lib/components/signDocumentButton.svelte';
 	const slotTypeCaster = createSlotTypeCaster<CrudExampleRecord>();
 </script>
 
@@ -36,7 +35,7 @@
 			}}
 			let:record
 		>
-		<SignDocumentButton {record} />
+		<!-- <SignDocumentButton {record} /> -->
 		</RecordsTable>
 	</RecordsManager>
 </div>
